FossilOrigin-Name: be0df0a65fc781fde6e94c9370b4819c703d3b270c707dd59e587a2a6f36c9f4

+++ ext/lsm1/lsm_varint.c
@@ -176,7 +176,7 @@ int lsmVarintGet32(u8 *z, int *piVal){
   }
 
   ret = lsmSqlite4GetVarint64(z, &i);
-  *piVal = i;
+  *piVal = (int)i;
   return ret;
 }
